Create README
Alternative Energy Resources Prediction - Wind, Solar & Nuclear

After realizing dangers of extreme climate change the need to reduce for greenhouse gas emission is increasing publically on vast scale in the world. Majority of these gases come from CO2 so the need for finding alternative energy resources are more than ever.

I performed analysis on different countries to find out the renewable energy they produced from 1990-2014. (Nuclear energy is not renewable but included for analysis purpose only).

Original data set is from "International Energy Statistics". A selected group of countries are used for analysis. We'll visualize the effort and success of different countries in producing wind, solar and nuclear energy resources.

The basic criteria to chose countries is total GDP, however because not countries from all continents are included, there is room to improve and add more countries for a wide spectrum of analysis.